Add alternate Models with Entities
