+++ src/Entities/Group.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Myth\Auth\Entities;
+
+use CodeIgniter\Entity\Entity;
+
+/**
+ * Group Entity
+ *
+ * As of version 1.2 this class is used by the new GroupModel
+ * to allow using a strongly-typed return. Any logic in this
+ * class should not be relied on within this library.
+ *
+ * @since 1.2.0
+ */
+class Group extends Entity
+{
+}
