+++ src/Entities/Permission.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Myth\Auth\Entities;
+
+use CodeIgniter\Entity\Entity;
+
+/**
+ * Permission Entity
+ *
+ * As of version 1.2 this class is used by the new PermissionModel
+ * to allow using a strongly-typed return. Any logic in this
+ * class should not be relied on within this library.
+ *
+ * @since 1.2.0
+ */
+class Permission extends Entity
+{
+}
